requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1

# Data processing
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.4
python-dateutil==2.9.0

# Google Sheets
google-auth==2.29.0
google-api-python-client==2.128.0
